% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tooltip.R
\name{add_tooltip}
\alias{add_tooltip}
\title{FUNCTION_TITLE}
\usage{
add_tooltip(nodes_and_edges)
}
\arguments{
\item{nodes_and_edges}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{group_by}},\code{\link[dplyr]{summarise}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{mutate_all}}
\code{\link[tidyr]{pivot_longer}},\code{\link[tidyr]{unite}}
\code{\link[tibble]{rownames}}
}
