% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{create_legend}
\alias{create_legend}
\title{FUNCTION_TITLE}
\usage{
create_legend(graph)
}
\arguments{
\item{graph}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{set_names}},\code{\link[purrr]{transpose}},\code{\link[purrr]{map}}
\code{\link[glue]{glue}}
\code{\link[huxtable]{huxtable}},\code{\link[huxtable]{background_color}},\code{\link[huxtable]{by_rows}},\code{\link[huxtable]{themes}}
\code{\link[dplyr]{select}}
}
