% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_omop}
\alias{fetch_omop}
\title{FUNCTION_TITLE}
\usage{
fetch_omop(
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  type_from = concept_class_id,
  label_glue = "{vocabulary_id}\\n{concept_class_id}\\n({standard_concept})\\n",
  schema = "omop_vocabulary",
  verbose = FALSE,
  render_sql = FALSE,
  version_key = get_version_key()
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{type_from}{PARAM_DESCRIPTION, Default: concept_class_id}

\item{schema}{PARAM_DESCRIPTION, Default: 'omop_vocabulary'}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{render_sql}{PARAM_DESCRIPTION, Default: FALSE}

\item{version_key}{PARAM_DESCRIPTION, Default: get_version_key()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
\code{\link[pg13]{query}}
\code{\link[dplyr]{arrange}},\code{\link[dplyr]{distinct}},\code{\link[dplyr]{select}},\code{\link[dplyr]{reexports}},\code{\link[dplyr]{mutate-joins}},\code{\link[dplyr]{bind}},\code{\link[dplyr]{rename}}
\code{\link[cli]{cli_progress_bar}},\code{\link[cli]{cli_abort}}
\code{\link[purrr]{transpose}},\code{\link[purrr]{map}},\code{\link[purrr]{reduce}}
}
