% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{get_version_key}
\alias{get_version_key}
\title{FUNCTION_TITLE}
\usage{
get_version_key(
  conn,
  conn_fun = "pg13::local_connect(verbose=FALSE)",
  log_schema = "public",
  log_table = "setup_athena_log",
  log_timestamp_field = "sa_datetime",
  template_only = FALSE,
  sql_only = FALSE
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION, Default: 'pg13::local_connect(verbose=FALSE)'}

\item{log_schema}{PARAM_DESCRIPTION, Default: 'public'}

\item{log_table}{PARAM_DESCRIPTION, Default: 'setup_athena_log'}

\item{log_timestamp_field}{PARAM_DESCRIPTION, Default: 'sa_datetime'}

\item{template_only}{PARAM_DESCRIPTION, Default: FALSE}

\item{sql_only}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
\code{\link[pg13]{query}}
}
