% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{hux_bold_values}
\alias{hux_bold_values}
\title{FUNCTION_TITLE}
\usage{
hux_bold_values(ht, ..., values, ignore_na = TRUE)
}
\arguments{
\item{ht}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{values}{PARAM_DESCRIPTION}

\item{ignore_na}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
