% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{hux_colorize}
\alias{hux_colorize}
\title{FUNCTION_TITLE}
\usage{
hux_colorize(ht, ...)
}
\arguments{
\item{ht}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[huxtable]{huxtable}},\code{\link[huxtable]{background_color}}
}
