% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{hux_pretty_percents}
\alias{hux_pretty_percents}
\title{FUNCTION_TITLE}
\usage{
hux_pretty_percents(ht, ..., digits = 1)
}
\arguments{
\item{ht}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{digits}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[huxtable]{huxtable}}
}
