% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammer.R
\name{map_edge_attributes}
\alias{map_edge_attributes}
\title{FUNCTION_TITLE}
\usage{
map_edge_attributes(
  nodes_and_edges,
  fontsize = 26,
  len = 5,
  label_from = relationship_name,
  style_from = NULL,
  style_map = NULL,
  penwidth_from = NULL,
  penwidth_map = NULL,
  color_from = NULL,
  color_map = NULL,
  arrowsize_from = NULL,
  arrowsize_map = NULL,
  arrowhead_from = is_hierarchical,
  arrowhead_map = c(`1` = "vee", `0` = "none"),
  arrowtail_from = NULL,
  arrowtail_map = NULL,
  fontname_from = NULL,
  fontname_map = NULL,
  fontsize_from = NULL,
  fontsize_map = NULL,
  fontcolor_from = NULL,
  fontcolor_map = NULL,
  tooltip_from = NULL,
  tooltip_map = NULL,
  URL_from = NULL,
  URL_map = NULL,
  edgetooltip_from = NULL,
  edgetooltip_map = NULL,
  edgeURL_from = NULL,
  edgeURL_map = NULL,
  dir_from = NULL,
  dir_map = NULL,
  headtooltip_from = NULL,
  headtooltip_map = NULL,
  headURL_from = NULL,
  headURL_map = NULL,
  headclip_from = NULL,
  headclip_map = NULL,
  headlabel_from = NULL,
  headlabel_map = NULL,
  headport_from = NULL,
  headport_map = NULL,
  tailtooltip_from = NULL,
  tailtooltip_map = NULL,
  tailURL_from = NULL,
  tailURL_map = NULL,
  tailclip_from = NULL,
  tailclip_map = NULL,
  taillabel_from = NULL,
  taillabel_map = NULL,
  tailport_from = NULL,
  tailport_map = NULL,
  decorate_from = NULL,
  decorate_map = NULL
)
}
\arguments{
\item{nodes_and_edges}{PARAM_DESCRIPTION}

\item{fontsize}{PARAM_DESCRIPTION, Default: 12}

\item{len}{PARAM_DESCRIPTION, Default: 1}

\item{label_from}{PARAM_DESCRIPTION, Default: relationship_name}

\item{style_from}{PARAM_DESCRIPTION, Default: NULL}

\item{style_map}{PARAM_DESCRIPTION, Default: NULL}

\item{penwidth_from}{PARAM_DESCRIPTION, Default: NULL}

\item{penwidth_map}{PARAM_DESCRIPTION, Default: NULL}

\item{color_from}{PARAM_DESCRIPTION, Default: NULL}

\item{color_map}{PARAM_DESCRIPTION, Default: NULL}

\item{arrowsize_from}{PARAM_DESCRIPTION, Default: NULL}

\item{arrowsize_map}{PARAM_DESCRIPTION, Default: NULL}

\item{arrowhead_from}{PARAM_DESCRIPTION, Default: is_hierarchical}

\item{arrowhead_map}{PARAM_DESCRIPTION, Default: c(\code{1} = "vee", \code{0} = "none")}

\item{arrowtail_from}{PARAM_DESCRIPTION, Default: NULL}

\item{arrowtail_map}{PARAM_DESCRIPTION, Default: NULL}

\item{fontname_from}{PARAM_DESCRIPTION, Default: NULL}

\item{fontname_map}{PARAM_DESCRIPTION, Default: NULL}

\item{fontsize_from}{PARAM_DESCRIPTION, Default: NULL}

\item{fontsize_map}{PARAM_DESCRIPTION, Default: NULL}

\item{fontcolor_from}{PARAM_DESCRIPTION, Default: NULL}

\item{fontcolor_map}{PARAM_DESCRIPTION, Default: NULL}

\item{tooltip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{tooltip_map}{PARAM_DESCRIPTION, Default: NULL}

\item{URL_from}{PARAM_DESCRIPTION, Default: NULL}

\item{URL_map}{PARAM_DESCRIPTION, Default: NULL}

\item{edgetooltip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{edgetooltip_map}{PARAM_DESCRIPTION, Default: NULL}

\item{edgeURL_from}{PARAM_DESCRIPTION, Default: NULL}

\item{edgeURL_map}{PARAM_DESCRIPTION, Default: NULL}

\item{dir_from}{PARAM_DESCRIPTION, Default: NULL}

\item{dir_map}{PARAM_DESCRIPTION, Default: NULL}

\item{headtooltip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{headtooltip_map}{PARAM_DESCRIPTION, Default: NULL}

\item{headURL_from}{PARAM_DESCRIPTION, Default: NULL}

\item{headURL_map}{PARAM_DESCRIPTION, Default: NULL}

\item{headclip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{headclip_map}{PARAM_DESCRIPTION, Default: NULL}

\item{headlabel_from}{PARAM_DESCRIPTION, Default: NULL}

\item{headlabel_map}{PARAM_DESCRIPTION, Default: NULL}

\item{headport_from}{PARAM_DESCRIPTION, Default: NULL}

\item{headport_map}{PARAM_DESCRIPTION, Default: NULL}

\item{tailtooltip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{tailtooltip_map}{PARAM_DESCRIPTION, Default: NULL}

\item{tailURL_from}{PARAM_DESCRIPTION, Default: NULL}

\item{tailURL_map}{PARAM_DESCRIPTION, Default: NULL}

\item{tailclip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{tailclip_map}{PARAM_DESCRIPTION, Default: NULL}

\item{taillabel_from}{PARAM_DESCRIPTION, Default: NULL}

\item{taillabel_map}{PARAM_DESCRIPTION, Default: NULL}

\item{tailport_from}{PARAM_DESCRIPTION, Default: NULL}

\item{tailport_map}{PARAM_DESCRIPTION, Default: NULL}

\item{decorate_from}{PARAM_DESCRIPTION, Default: NULL}

\item{decorate_map}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate}},\code{\link[dplyr]{distinct}}
}
