% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagrammer.R
\name{map_node_attributes}
\alias{map_node_attributes}
\title{FUNCTION_TITLE}
\usage{
map_node_attributes(
  nodes_and_edges,
  fontsize = 26,
  fontname = NULL,
  width = 4,
  height = 4,
  shape_from = standard_concept,
  shape_map = c(C = "square", S = "circle", `NA` = "circle"),
  shape_map_other = "circle",
  style_from = standard_concept,
  style_map = c(C = "filled", S = "filled", `NA` = "filled"),
  style_map_other = NULL,
  penwidth_from = NULL,
  penwidth_map = NULL,
  color_from = vocabulary_id,
  color_map = vocabulary_id_standard_colors,
  color_map_other = "gray20",
  fillcolor_from = vocabulary_id,
  fillcolor_map = vocabulary_id_standard_colors,
  fillcolor_map_other = "gray20",
  image_from = NULL,
  image_map = NULL,
  fontname_from = NULL,
  fontname_map = NULL,
  fontcolor_from = standard_concept,
  fontcolor_map = c(C = "black", S = "black", `NA` = "gray40"),
  fontcolor_map_other = NULL,
  peripheries_from = NULL,
  peripheries_map = NULL,
  x_from = NULL,
  x_map = NULL,
  y_from = NULL,
  y_map = NULL,
  group_from = NULL,
  group_map = NULL,
  xlabel_from = NULL,
  xlabel_map = NULL,
  URL_from = NULL,
  URL_map = NULL,
  sides_from = NULL,
  sides_map = NULL,
  orientation_from = NULL,
  orientation_map = NULL,
  skew_from = NULL,
  skew_map = NULL,
  distortion_from = NULL,
  distortion_map = NULL,
  gradientangle_from = NULL,
  gradientangle_map = NULL,
  fixedsize_from = NULL,
  fixedsize_map = NULL,
  labelloc_from = NULL,
  labelloc_map = NULL,
  margin_from = NULL,
  margin_map = NULL
)
}
\arguments{
\item{nodes_and_edges}{PARAM_DESCRIPTION}

\item{fontsize}{PARAM_DESCRIPTION, Default: 12}

\item{fontname}{PARAM_DESCRIPTION, Default: NULL}

\item{width}{PARAM_DESCRIPTION, Default: 1.5}

\item{height}{PARAM_DESCRIPTION, Default: 1.5}

\item{shape_from}{PARAM_DESCRIPTION, Default: standard_concept}

\item{shape_map}{PARAM_DESCRIPTION, Default: c(C = "square", S = "circle", \code{NA} = "circle")}

\item{shape_map_other}{PARAM_DESCRIPTION, Default: 'circle'}

\item{style_from}{PARAM_DESCRIPTION, Default: standard_concept}

\item{style_map}{PARAM_DESCRIPTION, Default: c(C = "filled", S = "filled", \code{NA} = "filled")}

\item{style_map_other}{PARAM_DESCRIPTION, Default: NULL}

\item{penwidth_from}{PARAM_DESCRIPTION, Default: NULL}

\item{penwidth_map}{PARAM_DESCRIPTION, Default: NULL}

\item{color_from}{PARAM_DESCRIPTION, Default: vocabulary_id}

\item{color_map}{PARAM_DESCRIPTION, Default: vocabulary_id_standard_colors}

\item{color_map_other}{PARAM_DESCRIPTION, Default: 'gray20'}

\item{fillcolor_from}{PARAM_DESCRIPTION, Default: vocabulary_id}

\item{fillcolor_map}{PARAM_DESCRIPTION, Default: vocabulary_id_standard_colors}

\item{fillcolor_map_other}{PARAM_DESCRIPTION, Default: 'gray20'}

\item{image_from}{PARAM_DESCRIPTION, Default: NULL}

\item{image_map}{PARAM_DESCRIPTION, Default: NULL}

\item{fontname_from}{PARAM_DESCRIPTION, Default: NULL}

\item{fontname_map}{PARAM_DESCRIPTION, Default: NULL}

\item{fontcolor_from}{PARAM_DESCRIPTION, Default: standard_concept}

\item{fontcolor_map}{PARAM_DESCRIPTION, Default: c(C = "black", S = "black", \code{NA} = "gray40")}

\item{fontcolor_map_other}{PARAM_DESCRIPTION, Default: NULL}

\item{peripheries_from}{PARAM_DESCRIPTION, Default: NULL}

\item{peripheries_map}{PARAM_DESCRIPTION, Default: NULL}

\item{x_from}{PARAM_DESCRIPTION, Default: NULL}

\item{x_map}{PARAM_DESCRIPTION, Default: NULL}

\item{y_from}{PARAM_DESCRIPTION, Default: NULL}

\item{y_map}{PARAM_DESCRIPTION, Default: NULL}

\item{group_from}{PARAM_DESCRIPTION, Default: NULL}

\item{group_map}{PARAM_DESCRIPTION, Default: NULL}

\item{xlabel_from}{PARAM_DESCRIPTION, Default: NULL}

\item{xlabel_map}{PARAM_DESCRIPTION, Default: NULL}

\item{URL_from}{PARAM_DESCRIPTION, Default: NULL}

\item{URL_map}{PARAM_DESCRIPTION, Default: NULL}

\item{sides_from}{PARAM_DESCRIPTION, Default: NULL}

\item{sides_map}{PARAM_DESCRIPTION, Default: NULL}

\item{orientation_from}{PARAM_DESCRIPTION, Default: NULL}

\item{orientation_map}{PARAM_DESCRIPTION, Default: NULL}

\item{skew_from}{PARAM_DESCRIPTION, Default: NULL}

\item{skew_map}{PARAM_DESCRIPTION, Default: NULL}

\item{distortion_from}{PARAM_DESCRIPTION, Default: NULL}

\item{distortion_map}{PARAM_DESCRIPTION, Default: NULL}

\item{gradientangle_from}{PARAM_DESCRIPTION, Default: NULL}

\item{gradientangle_map}{PARAM_DESCRIPTION, Default: NULL}

\item{fixedsize_from}{PARAM_DESCRIPTION, Default: NULL}

\item{fixedsize_map}{PARAM_DESCRIPTION, Default: NULL}

\item{labelloc_from}{PARAM_DESCRIPTION, Default: NULL}

\item{labelloc_map}{PARAM_DESCRIPTION, Default: NULL}

\item{margin_from}{PARAM_DESCRIPTION, Default: NULL}

\item{margin_map}{PARAM_DESCRIPTION, Default: NULL}

\item{tooltip_from}{PARAM_DESCRIPTION, Default: NULL}

\item{tooltip_map}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[cli]{cli_alert}}
\code{\link[dplyr]{tidyeval-compat}},\code{\link[dplyr]{mutate}}
\code{\link[purrr]{keep}}
\code{\link[tibble]{as_tibble}}
}
