% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_to_value}
\alias{map_to_value}
\title{FUNCTION_TITLE}
\usage{
map_to_value(x, map_assignment, other)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{map_assignment}{PARAM_DESCRIPTION}

\item{other}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_replace_na}}
\code{\link[cli]{cli_alert}}
\code{\link[forcats]{fct_recode}}
}
