% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_fetch_concept_classes.R
\name{ne_fetch_concept_classes}
\alias{ne_fetch_concept_classes}
\title{FUNCTION_TITLE}
\usage{
ne_fetch_concept_classes(
  conn,
  conn_fun,
  type_from = concept_class_id,
  label_glue = "{vocabulary_id}\\n{concept_class_id}\\n({standard_concept})\\n"
)
}
\arguments{
\item{conn}{PARAM_DESCRIPTION}

\item{conn_fun}{PARAM_DESCRIPTION}

\item{type_from}{PARAM_DESCRIPTION, Default: concept_class_id}

\item{label_glue}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[pg13]{query}}
\code{\link[R.cache]{loadCache}},\code{\link[R.cache]{saveCache}}
\code{\link[cli]{cli_progress_bar}}
\code{\link[tidyr]{extract}}
\code{\link[glue]{glue}}
}
