% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{read_sql_template}
\alias{read_sql_template}
\title{FUNCTION_TITLE}
\usage{
read_sql_template(file)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
