% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{save_to_cache}
\alias{save_to_cache}
\title{FUNCTION_TITLE}
\usage{
save_to_cache(resultset, sql, version_key)
}
\arguments{
\item{resultset}{PARAM_DESCRIPTION}

\item{sql}{PARAM_DESCRIPTION}

\item{version_key}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[R.cache]{saveCache}}
}
